chore: refactored proof-entry-input component into sub-components for UI elements
